test: Ass tests for the favorite counter in the item detail
