+++ webapp/src/utils/test.tsx
@@ -1,7 +1,9 @@
 import { render, RenderResult, waitFor } from '@testing-library/react'
-import TranslationProvider from 'decentraland-dapps/dist/providers/TranslationProvider'
+import { ConnectedRouter } from 'connected-react-router'
+import { createMemoryHistory } from 'history'
 import { Provider } from 'react-redux'
 import { Store } from 'redux'
+import TranslationProvider from 'decentraland-dapps/dist/providers/TranslationProvider'
 import { RootState } from '../modules/reducer'
 import { initTestStore } from '../modules/store'
 import * as locales from '../modules/translation/locales'
@@ -17,12 +19,13 @@ export function renderWithProviders(
       storage: { loading: false },
       translation: { data: locales, locale: 'en' }
     })
+  const history = createMemoryHistory()
 
   function AppProviders({ children }: { children: JSX.Element }) {
     return (
       <Provider store={initializedStore}>
         <TranslationProvider locales={Object.keys(locales)}>
-          {children}
+          <ConnectedRouter history={history}>{children}</ConnectedRouter>
         </TranslationProvider>
       </Provider>
     )
@@ -31,12 +34,11 @@ export function renderWithProviders(
   return render(component, { wrapper: AppProviders })
 }
 
-
 export async function waitForComponentToFinishLoading(screen: RenderResult) {
   // TODO: Make loader accessible so we can get the info without using the container ui#310
   await waitFor(() =>
-    expect(screen.container.getElementsByClassName('loader-container').length).toEqual(
-      0
-    )
+    expect(
+      screen.container.getElementsByClassName('loader-container').length
+    ).toEqual(0)
   )
 }
